Reorganize and add error handling
